% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampler.R
\name{box_ss}
\alias{box_ss}
\title{Simulate one or more sub-samples from a 100-cell subsampler}
\usage{
box_ss(ss = 10, N = 200, n_iter = 1e+05)
}
\arguments{
\item{ss}{The number cells that make up the subsample (equal to the %
subsample size)}

\item{N}{The total number of specimens in the sample}

\item{n_iter}{the number of repeated subsamples (with replacement) to be
simulated}
}
\value{
a vector of length n_iter of counts in the subsample
}
\description{
Simulate one or more sub-samples from a 100-cell subsampler
}
\details{
the function simulates the process of evenly spreading the sample
    across the 100 cells (allocating each of the N specimens to a cell drawn
    randomly from a uniform distribution between 1 and 100), and then
    randomly selecting ss cells to draw the subsample from.
}
\examples{
     box_ss(10,200,10)
}
