% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampler.R
\name{gamma_params_from_s_c}
\alias{gamma_params_from_s_c}
\title{Calculate Gamma distribution parameters for a sub-sample from a 100-cell subsampler}
\usage{
gamma_params_from_s_c(ss_proportion, count_in_subsample)
}
\arguments{
\item{ss_proportion}{The size of the subsample as a proportion of the whole sample}

\item{count_in_subsample}{The number of specimens collected in a subsample}
}
\value{
a list of 4 values
   \code{d_var} variance of the gamma distribution
   \code{mo} mode/mean, calculated as count_in_subsample/ss_proportion
   \code{mo_adj} adjusted mode/mean used for determining rate and shape
   \code{rate} the rate (beta) parameter of the distribution
   \code{shape} the shape (alpha) parameter of the distribution
}
\description{
Calculate Gamma distribution parameters for a sub-sample from a 100-cell subsampler
}
\details{
the mo estimate is based on the general relationship between mo and
   the combination of count_in_subsample and ss_proportion, derived by
   simulation. Shape and rate are estimated by moment method from d_var and
   mo_adj. See manuscript for more details.
}
\examples{
     dist_0_10 <- gamma_params_from_s_c(0.1,0)
     plot(0:40, dgamma(0:40, dist_0_10$shape, dist_0_10$rate),
          type = 'b',xlab = "T", ylab = "Density", las = 1)
     dist_10_25 <- gamma_params_from_s_c(0.25,10)
     plot(10:100, dgamma(10:100, dist_10_25$shape, dist_10_25$rate),
          type = 'l',xlab = "T", ylab = "Density", las = 1)
}
