% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampler.R
\name{rate_for_c_zero}
\alias{rate_for_c_zero}
\title{Calculate rate parameter of Gamma distribution for a sub-sample of proportion s with count = 0}
\usage{
rate_for_c_zero(s)
}
\arguments{
\item{s}{The subsample proportion}
}
\value{
The rate parameter of the gamma distribution
}
\description{
Calculate rate parameter of Gamma distribution for a sub-sample of proportion s with count = 0
}
\details{
Internal function for ss_count_gamma_params()
}
