% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampler.R
\name{T_var}
\alias{T_var}
\title{A generalised approximation of variance of the distribution of T|(s, c)}
\usage{
T_var(s, c)
}
\arguments{
\item{s}{The subsample proportion}

\item{c}{The number of specimens counted in a subsample}
}
\value{
The variance of the gamma distribution
}
\description{
A generalised approximation of variance of the distribution of T|(s, c)
}
\details{
Internal function for ss_count_gamma_params(). Coefficients used to
  calculate var_slope and var_intercept were determined as described by Walsh
  (2022) Measurement-error models improve prediction from subsampled data
  (See code chunk c_var_params, and Fig S1-2 in Appendix S1.
}
